Redone mortgage using service
I defined an interface for the service in types/app/home.ts
From there i implemented it as a class (tried to have the constructor
working instead of having the create method but didn't figure it out,
I'll need to have another look once i have more time)
You do not need to reference the ts files like it was done in tests.
I also put mortgage in a service folder since we might have a few.
Used camelCase as well

And no you still don't need the $scope.watch but i didn't remove it
this time, I'll explain later.