Making idiomatic angular/typescript
Moved rent/mortgage/house functions to their own
services.
More directives to core as they can be reused
(maybe change namespace of angular modules?) and prefix
them with waw (we are wizards)
HomeController is now only a dumb interface between the views
and the services.
Use ng-change in the template to recompute.
Add a few tests

There are a few function names that I don't like but couldn't
find something better.
There are also probably lots of things to improve on